chore: correct chat.py #3 - correct invalid files due to location confusion
